CHATBOT-RAG/Flask/requirements.txt
streamlit
google-generativeai
python-dotenv
PyPDF2
langchain
langchain-google-genai
langchain-community
faiss-cpu
pillow
flask
flask-cors
